Wait for all output to finish
It is possible for the first future to finish quicker than the others.
At which point calling shutdown-agents causes some pieces not to print.
A similiar thing can happen with the exit call on error.

Process all futures before calling shutdown-agents. Instead of
immediately exiting on error, save an exit code to use at the end.
Include an await to make sure the agent finishes processing all reports.